+++ doc/changelog.rst
@@ -72,6 +72,8 @@ Bug fixes
   integrators, wrong derivatives would be computed.
 - Fixed a bug in tendon wrapping around spheres. Before this fix, tendons that wrapped around spheres with an
   externally-placed :ref:`sidesite<spatial-geom-sidesite>` could jump inside the sphere instead of wrapping around it.
+- Fixed a bug that caused :at:`meshdir` and :at:`texturedir` to be overwritten during model
+  :ref:`attachment<meAttachment>.
 
 Python bindings
 ^^^^^^^^^^^^^^^